update DataTable.php
Currently we don't have any option for state save with pagination, we are not able to freeze state save if we go to some detail page from third page and if we come back from that detail page it again reloads the first page of pagination. That page state we can achieve through this.